+++ src/main/java/com/farcr/nomansland/common/event/CommonSetupEvents.java
@@ -347,7 +347,7 @@ public class CommonSetupEvents {
 
     @SubscribeEvent
     public static void onServerStarting(ServerStartingEvent event) {
-        MeetingPointCommand.applyPersistedOverride(event.getServer());
+        MeetingPointCommand.initializeMeetingPoint(event.getServer());
     }
 
     @SubscribeEvent